requirements.txt
pytest
python-telegram-bot
matplotlib
python-decouple
SQLAlchemy
mysqlclient